+++ examples/todos/packages/todos_data_source/lib/src/todos_data_source.dart
@@ -7,19 +7,19 @@ import 'package:todos_data_source/todos_data_source.dart';
 /// * U - Update
 /// * D - Delete
 abstract class TodosDataSource {
-  /// Create and return the newly created todo.
+  /// Create and return the newly created `todo`.
   Future<Todo> create(Todo todo);
 
   /// Return all todos.
   Future<List<Todo>> readAll();
 
-  /// Return a todo with the provided [id] if one exists.
+  /// Return a `todo` with the provided [id] if one exists.
   Future<Todo?> read(String id);
 
-  /// Update the todo with the provided [id] to match [todo] and
-  /// return the updated todo.
+  /// Update the `todo` with the provided [id] to match [todo] and
+  /// return the updated `todo`.
   Future<Todo> update(String id, Todo todo);
 
-  /// Delete the todo with the provided [id] if one exists.
+  /// Delete the `todo` with the provided [id] if one exists.
   Future<void> delete(String id);
 }