+++ packages/dart_frog/lib/dart_frog.dart
@@ -13,9 +13,10 @@ export 'src/_internal.dart'
         fromShelfMiddleware,
         requestLogger,
         serve;
+export 'src/body_parsers/body_parsers.dart' show FormData, UploadedFile;
 export 'src/create_static_file_handler.dart' show createStaticFileHandler;
 export 'src/handler.dart' show Handler;
 export 'src/hot_reload.dart' show hotReload;
 export 'src/http_method.dart' show HttpMethod;
-export 'src/middleware.dart' show Middleware, HandlerUse;
+export 'src/middleware.dart' show HandlerUse, Middleware;
 export 'src/provider.dart' show provider;