+++ packages/dart_frog_cli/test/src/commands/create/create_test.dart
@@ -69,27 +69,59 @@ void main() {
 
     test('throws UsageException when project directory is invalid.', () async {
       final directory = Directory.systemTemp.createTempSync();
-      when(() => argResults.rest).thenReturn([directory.path]);
+      when(() => argResults.rest).thenReturn([directory.path, directory.path]);
       when<dynamic>(() => argResults['project-name']).thenReturn(null);
       expect(command.run, throwsA(isA<UsageException>()));
     });
 
-    test('generates a project successfully.', () async {
+    test('generates a project successfully (defaults)', () async {
       final generatorHooks = _MockGeneratorHooks();
+      final directory = Directory.current.absolute.path;
       when(
         () => generatorHooks.postGen(
           vars: any(named: 'vars'),
           workingDirectory: any(named: 'workingDirectory'),
         ),
       ).thenAnswer((_) async {});
-      when<dynamic>(() => argResults['project-name']).thenReturn('example');
-      when(() => argResults.rest).thenReturn(['.']);
+      when(() => argResults.rest).thenReturn([directory]);
       when(
         () => generator.generate(any(), vars: any(named: 'vars')),
       ).thenAnswer((_) async => []);
       when(() => generator.hooks).thenReturn(generatorHooks);
       final exitCode = await command.run();
       expect(exitCode, equals(ExitCode.success.code));
+      verify(
+        () => generator.generate(
+          any(),
+          vars: {'name': 'dart_frog_cli', 'output_directory': directory},
+        ),
+      ).called(1);
+    });
+
+    test('generates a project successfully (custom)', () async {
+      final generatorHooks = _MockGeneratorHooks();
+      final directory = Directory.current.absolute.path;
+      const projectName = 'example';
+      when(
+        () => generatorHooks.postGen(
+          vars: any(named: 'vars'),
+          workingDirectory: any(named: 'workingDirectory'),
+        ),
+      ).thenAnswer((_) async {});
+      when<dynamic>(() => argResults['project-name']).thenReturn(projectName);
+      when(() => argResults.rest).thenReturn([directory]);
+      when(
+        () => generator.generate(any(), vars: any(named: 'vars')),
+      ).thenAnswer((_) async => []);
+      when(() => generator.hooks).thenReturn(generatorHooks);
+      final exitCode = await command.run();
+      expect(exitCode, equals(ExitCode.success.code));
+      verify(
+        () => generator.generate(
+          any(),
+          vars: {'name': projectName, 'output_directory': directory},
+        ),
+      ).called(1);
     });
   });
 }