+++ packages/dart_frog/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.3.4
+
+- feat: add support for `multipart/form-data` ([#551](https://github.com/VeryGoodOpenSource/dart_frog/pull/551))
+
+# 0.3.3
+
+- deps: upgrade to `Dart ">=2.19.0 <3.0.0"`
+- deps: upgrade to `very_good_analysis ^4.0.0`
+
 # 0.3.2
 
 - feat: cache `Request` and `Response` body